# internal/auth/model/visibility.go
/*
 * Otto user service
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model
// Visibility the model 'Visibility'
type Visibility string

// List of Visibility
const (
	PUBLIC Visibility = "public"
	PRIVATE Visibility = "private"
)